requirements.txt
django==3.2.6
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary==2.9.1
